P4A-614 Back when edit incoming and outgoing

+++ k9mail/src/main/java/com/fsck/k9/activity/setup/AccountSetupBasics.java
@@ -83,11 +83,13 @@ public class AccountSetupBasics extends PEpImporterActivity {
             ft.addToBackStack("AccountSetupIncomingFragment");
             String accountUuid = getIntent().getStringExtra(EXTRA_ACCOUNT);
             ft.add(R.id.account_setup_container, AccountSetupIncomingFragment.actionEditIncomingSettings(accountUuid)).commit();
+            accountSetupNavigator.setIsEditting(true);
         } else if (isEditingOutgoingSettings) {
             FragmentTransaction ft = getFragmentManager().beginTransaction();
             ft.addToBackStack("AccountSetupIncomingFragment");
             String accountUuid = getIntent().getStringExtra(EXTRA_ACCOUNT);
             ft.add(R.id.account_setup_container, AccountSetupOutgoingFragment.intentActionEditOutgoingSettings(accountUuid)).commit();
+            accountSetupNavigator.setIsEditting(true);
         }
         else if (savedInstanceState == null) {
             accountSetupBasicsFragment = new AccountSetupBasicsFragment();
@@ -179,7 +181,7 @@ public class AccountSetupBasics extends PEpImporterActivity {
     }
 
     private void goBack() {
-        if (accountSetupNavigator.shouldDeleteAccount()) {
+        if (accountSetupNavigator.shouldDeleteAccount() && !isEditingIncomingSettings && !isEditingOutgoingSettings) {
             deleteAccount();
         }
         accountSetupNavigator.goBack(this, getFragmentManager());
